Start work on Truffle implementation!

+++ src/main/kotlin/xyz/angm/lox/Interpreter.kt
@@ -14,7 +14,7 @@ class Interpreter : Expression.Visitor<Any?>, Statement.Visitor<Unit> {
     private val locals = HashMap<Expression, Int>()
 
     init {
-        globals.define("clock", LoxNativeFunction("clock", 0) { _, _ -> System.currentTimeMillis() })
+        globals.define("clock", LoxNativeFunction("clock", 0) { _, _ -> System.currentTimeMillis().toDouble() })
         globals.define("print", LoxNativeFunction("print", 1) { _, arguments -> print(stringify(arguments[0])) })
         globals.define("printLine", LoxNativeFunction("printLine", 1) { _, arguments -> println(stringify(arguments[0])) })
